fixup! refactor(sdk): Use a builder pattern to construct Timeline
